Restructure client event files
