+++ client/src/components/Navbar.js
@@ -17,7 +17,6 @@ function Navbar() {
                 <Link to = "/signup"><li>Sign up</li></Link>
                 <Link to = "/login"><li>Login</li></Link>
                 <Link to = "/contact"><li>Contact</li></Link>
-                <Link to = "/event-details"><li>Event Details</li></Link>
                 <Link to = "/basket"><div className='basketLogo'><img src = {basket} alt = "Basket"></img></div></Link>
             </ul>
         </nav>
